Move both tests into a single test file
